SpreadsheetLabelNameResolverTesting2 extracted SpreadsheetLabelNameResolverTesting
