+++ src/main/java/walkingkooka/spreadsheet/convert/SpreadsheetConverterContextTesting.java
@@ -19,12 +19,12 @@ package walkingkooka.spreadsheet.convert;
 
 import walkingkooka.spreadsheet.meta.SpreadsheetMetadataLoaderTesting2;
 import walkingkooka.spreadsheet.provider.SpreadsheetProviderContextTesting;
-import walkingkooka.spreadsheet.reference.SpreadsheetLabelNameResolverTesting;
+import walkingkooka.spreadsheet.reference.SpreadsheetLabelNameResolverTesting2;
 import walkingkooka.storage.convert.StorageConverterContextTesting2;
 import walkingkooka.tree.json.convert.JsonNodeConverterContextTesting;
 
 public interface SpreadsheetConverterContextTesting<C extends SpreadsheetConverterContext> extends JsonNodeConverterContextTesting<C>,
-    SpreadsheetLabelNameResolverTesting<C>,
+    SpreadsheetLabelNameResolverTesting2<C>,
     SpreadsheetMetadataLoaderTesting2<C>,
     StorageConverterContextTesting2<C>,
     SpreadsheetProviderContextTesting<C> {
